Добавляет логику открытия большой фотографии при нажатии на миниатюру
